+++ packages/carbon-graphs/CHANGELOG.md
@@ -3,6 +3,12 @@
 ## Unreleased
 
 * Changed
+  * Added deprecation warning for `reflow`.
+
+* Added
+  * Added `reflowMultipleDatasets` to allow users to send in multiple datasets per graph instance.
+
+* Changed
   * Fixed linter issues.
 
 ## 2.16.1 - (February 2, 2021)